+++ Resources/views/Admin/index.html.twig
@@ -1,4 +1,4 @@
-{{ render(controller('ZikulaAdminModule:Admin:adminheader')) }}
+{{ adminHeader() }}
 <h3>
 	<span class="fa fa-bar-chart"></span> {{ __('Informations')}}
 </h3>
@@ -26,4 +26,4 @@
 </div>    
 </div>
 </div>
-{{ render(controller('ZikulaAdminModule:Admin:adminfooter')) }}
+{{ adminFooter() }}
