+++ Resources/views/Admin/tools.html.twig
@@ -1,4 +1,4 @@
-{{ render(controller('ZikulaAdminModule:Admin:adminheader')) }}
+{{ adminHeader() }}
 <div class="container-fluid">
 	<h3>
 		<span class="fa fa-magic"></span> {{ __('Tools')}}
@@ -136,4 +136,4 @@
 		</div>
 	</div>
 </div>
-{{ render(controller('ZikulaAdminModule:Admin:adminfooter')) }}
+{{ adminFooter() }}
