+++ Resources/views/Classic/index.html.twig
@@ -1,0 +1,35 @@
+{% extends 'ZikulaIntercomModule:Base:user.html.twig' %}
+{% block header %}
+    {% set settings=settings %}
+    {{ parent() }}
+    {{ pageAddAsset('stylesheet', zasset('@ZikulaIntercomModule:css/inbox.css')) }}
+{% endblock %} 
+{% block title %}
+    {% set title_icon="<i class='fa fa-inbox'></i>" %}
+    {% set title=__(box) %}
+    {{ parent() }}
+{% endblock %}
+{% block content %}
+    <div id="intercom_inbox" class="col-md-12 clearfix">   
+            {{ include('ZikulaIntercomModule:Inbox:list.header.html.twig') }}
+            {{ include('ZikulaIntercomModule:Inbox:list.selected.html.twig') }}
+            {#render#}
+            <div id="conversations" class="conversations-list list-group clearfix"> 
+             {{ include('ZikulaIntercomModule:Inbox:list.html.twig') }}
+            </div>    
+            {{ include('ZikulaIntercomModule:Inbox:list.footer.html.twig') }}
+    </div>
+{% endblock %}
+{% block footer %}
+    {{ parent() }}
+    {{ include('ZikulaIntercomModule:Base:modal.html.twig') }}
+    {{ pageAddAsset('javascript', zasset('@ZikulaIntercomModule:js/Zikula.Intercom.inbox.manager.js')) }}
+    <script>
+    (function($) {
+            $(function(){
+                Zikula.Intercom.InboxManager.init();  
+            });
+    })(jQuery);
+    </script>
+    {#{ dump() }#}
+{% endblock %}
