+++ Resources/views/Classic/list.conversation.first.html.twig
@@ -1,0 +1,1 @@
+<div class="conversation-first"> {{ message.text }} </div>
