+++ Resources/views/Classic/list.conversation.messages.html.twig
@@ -1,0 +1,8 @@
+<div class="conversation-list row">
+    <div class="col-md-12">
+            {% for messagex in message.conversation %}
+                {{ include('ZikulaIntercomModule:Inbox:list.message.single.html.twig', {'message': messagex,'align': messagex.sender == message.sender ? '' : 'pull-right' }) }}
+                {% else %}             
+                {% endfor %}
+    </div>
+</div>
