+++ Resources/views/Classic/list.conversation.statuses.html.twig
@@ -1,0 +1,15 @@
+<div class="btn-group btn-group-sm text-right pull-right">
+    {% if message.seen %}
+        <a class="btn text-muted" title="{{ __('Message seen')}} {{ message.seen|date("m/d/Y") }}"> <i class="fa fa-eye"> </i></a>
+    {% else %}
+        <a class="btn text-warning" title="{{ __('Unread')}}"> <i class="fa fa-eye-slash"> </i></a>
+    {% endif %}     
+    {% if message.storedbyrecipient %}
+        <a class="btn text-success" title="{{ __('Message is saved')}}"><i class='fa fa-save'></i> </a>  
+    {% else %}
+        <a class="btn text-muted" title="{{ __('Message is not saved')}}"><i class='fa fa-save'></i> </a> 
+    {% endif %}   
+
+    <div class="btn"><input class="conversation-select" data-id="{{ message.id }}" name="selectable[{{ message.id }}]" type="checkbox"></div>
+    <a data-toggle="collapse" data-seen="{{ message.seen is empty ? "false" : "true" }}" data-id="{{ message.id }}" class="conversation-details" data-target="#conversation-body-{{ message.id }}" > <span class="fa fa-angle-down fa-2x" title="{gt text='Read'}"> </span></a>   
+</div>
