+++ Resources/views/Classic/list.html.twig
@@ -1,0 +1,5 @@
+{% for message in messages %}
+    {{ include('ZikulaIntercomModule:Inbox:list.conversation.html.twig') }}
+{% else %}
+    <div class="list-group-item col-md-12 alert alert-info">{{ __('You currently have no messages.') }}</div>               
+{% endfor %}
