+++ Resources/views/Classic/list.selected.html.twig
@@ -1,0 +1,11 @@
+<div class="list-selected list-group-item clearfix hide" style="background: #E1FFB1;">
+    <div class="col-md-12">
+        {{ __('You have selected') }} <span class="selected-messages-count">0</span> {{ __('messages') }}                     
+        <div class="multi-actions btn-group btn-group-sm">  
+            <button type="button" data-action="delete-selected" class="multi-delete-selected btn text-danger" title="{{ __('Delete selected') }}"> {{ __('Delete selected') }} <i class='fa fa-trash'></i> </button>
+            <button type="button" data-action="markread-selected" class="multi-read-selected btn text-default" title="{{ __('Mark selected as read') }}">{{ __('Mark selected as read') }} <i class='fa fa-eye'></i> </button>
+            <button type="button" data-action="save-selected" class="multi-save-selected btn text-success" title="{{ __('Save selected') }}">{{ __('Save selected') }} <i class='fa fa-save'></i> </button>
+            <button type="button" data-action="unselect-selected" class="multi-unselect btn text-default" title="{{ __('Un select') }}">{{ __('Un select') }} <i class='fa fa-square'></i> </button>
+        </div>
+    </div>
+</div>
